Send `VoiceMessage` on becoming `ActiveSpeaker`
